+++ src/app/account/_components/step1.tsx
@@ -96,7 +96,7 @@ export default function SignupStep1() {
           {response?.message ? (
             <div className={`${form.username && response?.no ? "text-red-500" : response?.yes ? "text-green-400" : "text-gray-300"}`}>{response?.message}</div>
           ) : (
-            <div className="flex items-center gap-3">
+            <div className="flex items-center gap-3 ">
               <AiOutlineLoading3Quarters className="animate-spin" />
               <div className="animate-pulse">Checking</div>
             </div>
